third_party/WebKit/LayoutTests/http/tests/serviceworker/chromium.full-code-cache-install.html

<!DOCTYPE html>
<!-- This test requires internals because it tests Chromium's
  internal UseCounter. See https://crbug.com/768705. -->
<title>Service Worker: Full code cache install must be triggered by Origin-Trial</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
promise_test(t => {
    const disabled_url =
      'resources/full-code-cache-install-worker.php?';
    const disabled_scope =
      'resources/blank.html?full-code-cache-install-worker-disabled';
    const enabled_url =
      'resources/full-code-cache-install-worker.php?origin_trial=true';
    const enabled_scope =
      'resources/blank.html?full-code-cache-install-worker-enabled';
    // from web_feature.mojom
    const kScriptInCacheStorageInstalledWithMetadata = 2237;

    return service_worker_unregister_and_register(t,
                                                  disabled_url,
                                                  disabled_scope)
      .then(r => {
          add_completion_callback(() => { r.unregister(); });
          return wait_for_state(t, r.installing, 'activated');
        })
      .then(() => { return with_iframe(disabled_scope); })
      .then(frame => {
          add_completion_callback(() => { frame.remove(); });
          assert_false(
              frame.contentWindow.internals.isUseCounted(
                  frame.contentDocument,
                  kScriptInCacheStorageInstalledWithMetadata),
              'The usecounter must not be counted yet.');
          return service_worker_unregister_and_register(t,
                                                        enabled_url,
                                                        enabled_scope)
        })
      .then(r => {
        add_completion_callback(() => { r.unregister(); });
        return wait_for_state(t, r.installing, 'activated');
      })
      .then(() => { return with_iframe(enabled_scope); })
      .then(frame => {
          add_completion_callback(() => { frame.remove(); });
          assert_true(
              frame.contentWindow.internals.isUseCounted(
                  frame.contentDocument,
                  kScriptInCacheStorageInstalledWithMetadata),
              'The usecounter must be counted.');
        });
  }, 'Test full code cache install must be triggered by Origin-Trial');
</script>
